refactor: moved card & deck class to utils directory
