+++ src/utils/card.model.js
